requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pandas==1.5.3
numpy==1.23.5
faiss-cpu==1.7.3
sentence-transformers==2.2.2
transformers==4.26.1
torch==1.13.1
pydantic==1.10.7
streamlit==1.25.0
requests==2.28.2
python-multipart==0.0.5
prometheus-client==0.16.0
pytest==7.2.0
httpx==0.23.3
nltk==3.8.1
sqlalchemy==2.0.4
psycopg2-binary==2.9.5
alembic==1.10.2
gunicorn==20.1.0
plotly==5.13.1
matplotlib==3.7.1
pytest-cov==4.1.0
black==23.1.0
isort==5.12.0
pytest-mock==3.10.0
freezegun==1.2.2
python-dotenv==1.0.0
tenacity==8.2.2